requirement.txt
Flask
keras
tensorflow
numpy
matplotlib
scikit-image
Pillow
